Refactor notion lib
